winner searcher.service. and board-handler-service and types excluded to additional files. Bug fixed - left bottom diagonal winners found correctly
